refactor: Move Datum to charts
