+++ src/charts/ScatterChart.ts
@@ -1,6 +1,7 @@
 import { ScaleLinear, scaleLinear } from "d3-scale";
+import { Datum } from ".";
 import { ColorMap } from "../colors";
-import { Datum, Svg } from "../components";
+import { Svg } from "../components";
 import { BAR, getPathData } from "../coords";
 import { Dimensions, ResolvedDimensions } from "../dims";
 import {
