+++ src/charts/index.ts
@@ -1,5 +1,6 @@
 export * as BarChart from "./BarChart";
 export * as BubbleChart from "./BubbleChart";
+export * as Datum from "./Datum";
 export * as PieChart from "./PieChart";
 export * as ScatterChart from "./ScatterChart";
 export * as TreemapChart from "./TreemapChart";
