+++ src/components/index.ts
@@ -1,7 +1,6 @@
 export * as Axis from "./Axis";
 export * as AxisTick from "./AxisTick";
 export * as ColorLegend from "./ColorLegend";
-export * as Datum from "./Datum";
 export * as Step from "./Step";
 export * from "./Svg";
 export * as Text from "./Text";
